Users

+++ eco-friendly/src/components/Home/Home.js
@@ -1,15 +1,24 @@
-import React, * as react from 'react';
+import React, { Component } from 'react';
 import firebase from 'firebase/app';
 import 'firebase/auth';
 import connection from '../../helpers/data/connections';
+import Auth from '../Auth/Auth';
+import authRequests from '../../helpers/data/authRequests';
+import MyNavbar from '../MyNavbar/MyNavbar';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import Categories from '../pages/Categories/Categories';
 import ecoPointsRequest from '../../helpers/data/ecopointsRequest';
 import AddForm from '../pages/AddForm/AddForm';
+import EcoUser from '../pages/Users/User';
+import ecousersRequest from '../../helpers/data/ecousersRequests';
 
-class App extends react.Component {
+class App extends Component {
  state = {
    authed: false,
+   ecouser: {
+     userName: '',
+     points: 0,
+   },
    ecopoints: [],
    isEditing: false,
    editId: '-1',
@@ -22,25 +31,51 @@ class App extends react.Component {
    });
  }
 
- componentDidMount() {
-   connection();
-   ecoPointsRequest.getRequest()
+ getEcouser = () => {
+   const currentUid = authRequests.getCurrentUid();
+   ecousersRequest.getEcoUserByUid(currentUid)
+     .then((ecouser) => {
+       this.setState({ ecouser });
+     })
+     .catch((error) => {
+       console.error('error on users', error);
+     });
+ }
+
+ getEcoPoints = () => {
+   const currentUid = authRequests.getCurrentUid();
+   ecoPointsRequest.getRequest(currentUid)
      .then((ecopoints) => {
        this.setState({ ecopoints });
      })
      .catch(err => console.error('error with ecopoint GET', err));
+ }
+
+ // Add UID
+ componentDidMount() {
+   connection();
 
    this.removeListener = firebase.auth().onAuthStateChanged((user) => {
      if (user) {
        this.setState({
          authed: true,
        });
+       this.getEcoPoints();
+       this.getEcouser();
      } else {
        this.setState({
          authed: false,
        });
      }
    });
+ }
+
+ componentWillUnmount() {
+   this.removeListener();
+ }
+
+ isAuthenticated = () => {
+   this.setState({ authed: true });
  }
 
  deleteOne = (ecopointId) => {
@@ -53,8 +88,6 @@ class App extends react.Component {
      })
      .catch(err => console.error('error with delete single', err));
  }
-
- passEcopointToEdit = ecopointId => this.setState({ isEditing: true, editId: ecopointId });
 
 formSubmitEvent = (newEcopoint) => {
   const { isEditing, editId } = this.state;
@@ -70,6 +103,12 @@ formSubmitEvent = (newEcopoint) => {
   } else {
     ecoPointsRequest.postRequest(newEcopoint)
       .then(() => {
+        const pointTotal = newEcopoint.points + this.state.ecouser.points;
+        ecousersRequest.updateEcoUser(this.state.ecouser.id, { points: pointTotal })
+          .then(() => {
+            this.getEcouser();
+          })
+          .catch(err => console.error('error with ecopoints post', err));
         ecoPointsRequest.getRequest()
           .then((ecopoints) => {
             this.setState({ ecopoints });
@@ -79,28 +118,51 @@ formSubmitEvent = (newEcopoint) => {
   }
 }
 
+passEcopointToEdit = ecopointId => this.setState({ isEditing: true, editId: ecopointId });
+
 render() {
-  const passEcopointToEdit = (ecopointId) => {
-    this.setState({ isEditing: true, editId: ecopointId });
-  };
   const {
+    authed,
     ecopoints,
     isEditing,
     editId,
   } = this.state;
+
+  const logoutClickEvent = () => {
+    authRequests.logoutUser();
+    this.setState({ authed: false });
+  };
+
+  if (!authed) {
+    return (
+      <div className="App">
+        <MyNavbar isAuthed={authed} logoutClickEvent={logoutClickEvent} />
+        <div className="row">
+          <Auth isAuthenticated={this.isAuthenticated}/>
+        </div>
+      </div>
+    );
+  }
   return (
-    <div className="Home">
+    <div className="App">
+      <MyNavbar isAuthed={authed} logoutClickEvent={logoutClickEvent} />
+      <div className="row">
         <Categories
         ecopoints={ecopoints}
         deleteSingleEcopoint={this.deleteOne}
         passEcopointToEdit={this.passEcopointToEdit}
         onListingSelection={this.ecopointSelectEvent}
         />
-      ));
-       <div className="row">
-        <AddForm onSubmit={this.formSubmitEvent} isEditing={isEditing} editId={editId}/>
-        </div>
       </div>
+      <div className="row">
+        <AddForm onSubmit={this.formSubmitEvent} isEditing={isEditing} editId={editId}/>/>
+      </div>
+    <div className="row">
+    <EcoUser
+    ecouser={this.state.ecouser}
+    />
+    </div>
+    </div>
   );
 }
 }
